collect packets based on time
this change switches to harvesting the count packets per time increment
instead of per packet received. hopefully this give a more accurate view
of the messages occuring as they are not bound as tightly to the message
count read limit in zaqar.

also modifying the graph to use bars instead of lines.