refactor log line behavior
this change reformats the way log lines and count packets are stored in
the database. the log lines are now stored as individual documents
including their service and the log line, and the count packets contain
references to those line records. a new database collection for the log
lines is also added.

* change caravan_master normalize function to split apart the log lines
  and create individual ids for each line.
* change caravan_master store function to store the count packets and
  the log line packets.
* change shiny_squirrel sorted logs view function to handle new
  individual log line records.